requirements.txt
fastapi
uvicorn
python-dotenv
asyncpg
psycopg2
SQLAlchemy
alembic
passlib[bcrypt]
pyjwt
python-jose
pydantic
python-multipart
httpx